requirements.txt
langchain
langchain-groq
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
python-docx
python-pptx
PyMuPDF
python-dotenv
tqdm
docx2txt